@@ -16,61 +16,61 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^12.2.12",
-    "@angular/cdk": "^12.2.0",
-    "@angular/common": "^12.2.12",
-    "@angular/compiler": "^12.2.11",
-    "@angular/core": "^12.2.11",
-    "@angular/forms": "^12.2.12",
-    "@angular/material": "^12.2.12",
-    "@angular/platform-browser": "^12.2.12",
-    "@angular/platform-browser-dynamic": "^12.2.12",
-    "@angular/router": "^12.2.12",
+    "@angular/animations": "^13.0.0",
+    "@angular/cdk": "^13.0.0",
+    "@angular/common": "^13.0.0",
+    "@angular/compiler": "^13.0.0",
+    "@angular/core": "^13.0.0",
+    "@angular/forms": "^13.0.0",
+    "@angular/material": "^13.0.0",
+    "@angular/platform-browser": "^13.0.0",
+    "@angular/platform-browser-dynamic": "^13.0.0",
+    "@angular/router": "^13.0.0",
     "autoprefixer": "^10.4.0",
     "ng-click-outside": "^9.0.0",
-    "ngx-cookie-service": "^12.0.3",
+    "ngx-cookie-service": "^13.0.0",
     "ngx-mask": "^12.0.0",
-    "rxjs": "^6.6.7",
+    "rxjs": "^7.4.0",
     "tslib": "^2.3.1",
     "uuid": "^8.3.2",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1202.10",
-    "@angular-devkit/build-angular": "~12.2.12",
-    "@angular-eslint/builder": "^12.6.1",
-    "@angular-eslint/eslint-plugin": "^12.6.1",
-    "@angular-eslint/eslint-plugin-template": "^12.6.1",
-    "@angular-eslint/schematics": "^12.6.1",
-    "@angular-eslint/template-parser": "^12.6.1",
-    "@angular/cli": "^12.2.12",
-    "@angular/compiler-cli": "^12.2.12",
+    "@angular-devkit/architect": "^0.1300.1",
+    "@angular-devkit/build-angular": "^13.0.1",
+    "@angular-eslint/builder": "^13.0.0-alpha.0",
+    "@angular-eslint/eslint-plugin": "^13.0.0-alpha.0",
+    "@angular-eslint/eslint-plugin-template": "^13.0.0-alpha.0",
+    "@angular-eslint/schematics": "^13.0.0-alpha.0",
+    "@angular-eslint/template-parser": "^13.0.0-alpha.0",
+    "@angular/cli": "^13.0.1",
+    "@angular/compiler-cli": "^13.0.0",
     "@compodoc/compodoc": "^1.1.15",
-    "@strut-software/eslint-config": "^1.5.0",
+    "@strut-software/eslint-config": "^1.6.0",
     "@strut-software/stylelint-config": "^1.1.1",
-    "@types/css-font-loading-module": "0.0.6",
     "@types/jasmine": "^3.10.2",
-    "@types/node": "^16.11.6",
+    "@types/node": "^16.11.7",
     "@types/uuid": "^8.3.1",
     "@typescript-eslint/eslint-plugin": "4.33.0",
     "@typescript-eslint/parser": "4.33.0",
     "cross-env": "^7.0.3",
     "eslint": "^7.32.0",
     "eslint-plugin-angular": "^4.1.0",
+    "eslint-plugin-deprecation": "^1.2.1",
     "eslint-plugin-jsdoc": "^34.8.2",
     "eslint-plugin-rxjs": "^3.3.7",
     "eslint-plugin-rxjs-angular": "^1.0.6",
     "husky": "^7.0.4",
     "jasmine-core": "~3.10.1",
     "jasmine-spec-reporter": "~7.0.0",
-    "karma": "~6.3.7",
+    "karma": "~6.3.8",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.7.0",
     "ng-mocks": "^12.5.0",
     "pa11y-ci": "2.4.2",
-    "postcss-cli": "^9.0.1",
+    "postcss-cli": "^9.0.2",
     "protractor": "~7.0.0",
     "sassdoc": "^2.7.3",
     "stylelint": "^13.13.1",
@@ -78,6 +78,6 @@
     "stylelint-scss": "^3.21.0",
     "tailwindcss": "^2.2.19",
     "ts-node": "~10.4.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.4.4"
   }
 }
