@@ -1,7 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { first } from 'rxjs/operators';
-import { StationService } from '../../core/station.service';
-import { ErrorService } from '../../core/error.service';
+import { StationService } from 'src/app/core/station.service';
+import { ErrorService } from 'src/app/core/error.service';
+import { SidenavDrawerService } from 'src/app/core/sidenav-drawer.service';
 
 /**
  * Component for document drawer.
@@ -19,8 +20,11 @@ export class DocumentInfoDrawerComponent implements OnInit {
   /** The station rithmId. */
   stationRithmId = '';
 
-  constructor(private stationService: StationService,
-    private errorService: ErrorService) {
+  constructor(
+    private stationService: StationService,
+    private errorService: ErrorService,
+    private sidenavDrawerService: SidenavDrawerService,
+  ) {
 
   }
 
@@ -39,15 +43,17 @@ export class DocumentInfoDrawerComponent implements OnInit {
   getStatusDocumentEditable(stationRithmId: string): void {
     this.stationService.getStatusDocumentEditable(stationRithmId)
       .pipe(first())
-      .subscribe((documentEditableStatus) => {
-        if (documentEditableStatus) {
-          this.documentNameEditable = documentEditableStatus;
+      .subscribe({
+        next: (documentEditableStatus) => {
+          if (documentEditableStatus) {
+            this.documentNameEditable = documentEditableStatus;
+          }
+        }, error: (error: unknown) => {
+          this.errorService.displayError(
+            'Something went wrong on our end and we\'re looking into it. Please try again in a little while.',
+            error
+          );
         }
-      }, (error: unknown) => {
-        this.errorService.displayError(
-          'Something went wrong on our end and we\'re looking into it. Please try again in a little while.',
-          error
-        );
       });
   }
 
@@ -60,15 +66,26 @@ export class DocumentInfoDrawerComponent implements OnInit {
   updateStatusDocumentEditable(stationRithmId: string, newStatus: boolean): void {
     this.stationService.updateStatusDocumentEditable(stationRithmId, newStatus)
       .pipe(first())
-      .subscribe((documentEditableStatus) => {
-        if (documentEditableStatus) {
-          this.documentNameEditable = documentEditableStatus;
+      .subscribe({
+        next: (documentEditableStatus) => {
+          if (documentEditableStatus) {
+            this.documentNameEditable = documentEditableStatus;
+          }
+        }, error: (error: unknown) => {
+          this.errorService.displayError(
+            'Something went wrong on our end and we\'re looking into it. Please try again in a little while.',
+            error
+          );
         }
-      }, (error: unknown) => {
-        this.errorService.displayError(
-          'Something went wrong on our end and we\'re looking into it. Please try again in a little while.',
-          error
-        );
       });
+  }
+
+  /**
+   * Toggles the open state of the drawer for document info.
+   *
+   * @param drawerItem The drawer item to toggle.
+   */
+  toggleDrawer(drawerItem: 'documentInfo'): void {
+    this.sidenavDrawerService.toggleDrawer(drawerItem);
   }
 }
